Run JSDOM tests
This adds JSDOM to run through all of the same tests. This verifies the
polyfill works with both.
