+++ test/run-tests.js
@@ -1,0 +1,9 @@
+var testCharacterData = require("./test-characterData");
+var testChildList = require("./test-childList");
+var testAttributes = require("./test-attributes");
+
+module.exports = function(implName, window){
+	testCharacterData(implName, window);
+	testChildList(implName, window);
+	testAttributes(implName, window);
+};
