+++ test/test-can-vdom.html
@@ -1,0 +1,5 @@
+<!doctype html>
+<title>done-mutation-observer can-vdom</title>
+<script src="../node_modules/steal/steal.js"
+	data-main="~/test/test-can-vdom"></script>
+<div id="qunit-fixture"></div>
