+++ test/test-can-vdom.js
@@ -1,0 +1,4 @@
+var makeWindow = require("can-vdom/make-window/make-window");
+var runTests = require("./run-tests");
+
+runTests("can-vdom", makeWindow({}));
