+++ test/test-jsdom.html
@@ -1,11 +1,11 @@
 <!doctype html>
-<title>done-mutation-observer</title>
-<script src="node_modules/steal/steal.js"></script>
+<title>done-mutation-observer jsdom</title>
+<script src="../node_modules/steal/steal.js"></script>
 <script type="steal-module">
 	process.nextTick = function(cb) {
 		Promise.resolve().then(cb);
 	};
 	window.Buffer = require("buffer").Buffer;
-	steal.import("done-mutation-observer/done-mutation-observer-test");
+	steal.import("~/test/test-jsdom");
 </script>
 <div id="qunit-fixture"></div>
