+++ test/test-jsdom.js
@@ -1,0 +1,4 @@
+var JSDOM = require("jsdom").JSDOM;
+var runTests = require("./run-tests");
+
+runTests("jsdom", new JSDOM().window);
